requirements.txt
groq
python-dotenv
fastapi
uvicorn[standard]
pydantic
requests
gunicorn
